Simplify turn policy and turn storage
