+++ src/mindroom/bot.py
@@ -102,11 +102,6 @@ from .delivery_gateway import (
 from .delivery_gateway import (
     SuppressedPlaceholderCleanupError as _SuppressedPlaceholderCleanupError,
 )
-from .dispatch_planner import (
-    DispatchPlanner,
-    DispatchPlannerDeps,
-    IngressHookRunner,
-)
 from .edit_regenerator import EditRegenerator, EditRegeneratorDeps
 from .handled_turns import HandledTurnLedger, HandledTurnState
 from .inbound_turn_normalizer import (
@@ -143,6 +138,12 @@ from .scheduling import (
     restore_scheduled_tasks,
 )
 from .turn_controller import TurnController, TurnControllerDeps
+from .turn_policy import (
+    IngressHookRunner,
+    TurnPolicy,
+    TurnPolicyDeps,
+)
+from .turn_store import TurnStore, TurnStoreDeps
 
 if TYPE_CHECKING:
     from collections.abc import Awaitable, Callable, Sequence
@@ -324,12 +325,13 @@ class AgentBot:
     _runtime_view: BotRuntimeState
     _coalescing_gate: CoalescingGate
     _inbound_turn_normalizer: InboundTurnNormalizer
-    _dispatch_planner: DispatchPlanner
+    _turn_policy: TurnPolicy
     _conversation_resolver: ConversationResolver
     _conversation_state_writer: ConversationStateWriter
     _conversation_access: MatrixConversationAccess
     _delivery_gateway: DeliveryGateway
     _response_runner: ResponseRunner
+    _turn_store: TurnStore
     _tool_runtime_support: ToolRuntimeSupport
     _post_response_effects_support: PostResponseEffectsSupport
     _ingress_hook_runner: IngressHookRunner
@@ -457,6 +459,15 @@ class AgentBot:
             resolver=self._conversation_resolver,
             hook_context=self._hook_context_support,
         )
+        self._turn_store = TurnStore(
+            TurnStoreDeps(
+                agent_name=self.agent_name,
+                handled_turn_ledger=self.handled_turn_ledger,
+                state_writer=self._conversation_state_writer,
+                resolver=self._conversation_resolver,
+                tool_runtime=self._tool_runtime_support,
+            ),
+        )
         self._post_response_effects_support = PostResponseEffectsSupport(
             runtime=self._runtime_view,
             logger=self.logger,
@@ -490,29 +501,19 @@ class AgentBot:
                 get_logger=lambda: self.logger,
                 runtime_paths=self.runtime_paths,
                 agent_name=self.agent_name,
-                get_handled_turn_ledger=lambda: self.handled_turn_ledger,
                 resolver=self._conversation_resolver,
-                state_writer=self._conversation_state_writer,
-                tool_runtime=self._tool_runtime_support,
-                dispatch_hook_service=self._ingress_hook_runner,
+                turn_store=self._turn_store,
+                ingress_hook_runner=self._ingress_hook_runner,
                 generate_response=lambda **kwargs: self._generate_response(**kwargs),
             ),
         )
-        self._dispatch_planner = DispatchPlanner(
-            DispatchPlannerDeps(
+        self._turn_policy = TurnPolicy(
+            TurnPolicyDeps(
                 runtime=self._runtime_view,
                 logger=self.logger,
-                handled_turn_ledger=self.handled_turn_ledger,
                 runtime_paths=self.runtime_paths,
-                storage_path=self.storage_path,
                 agent_name=self.agent_name,
                 matrix_id=runtime_matrix_id,
-                normalizer=self._inbound_turn_normalizer,
-                resolver=self._conversation_resolver,
-                delivery_gateway=self._delivery_gateway,
-                response_runner=self._response_runner,
-                hook_service=self._ingress_hook_runner,
-                tool_runtime=self._tool_runtime_support,
             ),
         )
         self._turn_controller = TurnController(
@@ -520,16 +521,19 @@ class AgentBot:
                 runtime=self._runtime_view,
                 logger=self.logger,
                 runtime_paths=self.runtime_paths,
+                storage_path=self.storage_path,
                 agent_name=self.agent_name,
                 matrix_id=runtime_matrix_id,
-                handled_turn_ledger=self.handled_turn_ledger,
                 conversation_access=self._conversation_access,
                 resolver=self._conversation_resolver,
                 normalizer=self._inbound_turn_normalizer,
-                dispatch_planner=self._dispatch_planner,
+                turn_policy=self._turn_policy,
+                ingress_hook_runner=self._ingress_hook_runner,
                 response_runner=self._response_runner,
                 delivery_gateway=self._delivery_gateway,
                 state_writer=self._conversation_state_writer,
+                tool_runtime=self._tool_runtime_support,
+                turn_store=self._turn_store,
                 coalescing_gate=self._coalescing_gate,
                 edit_regenerator=self._edit_regenerator,
             ),
@@ -1313,7 +1317,7 @@ class AgentBot:
             self.logger.debug("ignoring_reaction_from_unauthorized_sender", user_id=event.sender)
             return
 
-        if not self._dispatch_planner.can_reply_to_sender(event.sender):
+        if not self._turn_policy.can_reply_to_sender(event.sender):
             self.logger.debug("Ignoring reaction due to reply permissions", sender=event.sender)
             return
 
@@ -1755,7 +1759,7 @@ class TeamBot(AgentBot):
         )
 
         configured_mode = TeamMode.COORDINATE if self.team_mode == "coordinate" else TeamMode.COLLABORATE
-        materializable_agent_names = self._dispatch_planner.materializable_agent_names()
+        materializable_agent_names = self._turn_policy.materializable_agent_names()
         team_resolution = resolve_configured_team(
             self.agent_name,
             self.team_agents,
