+++ src/mindroom/edit_regenerator.py
@@ -5,20 +5,12 @@ from __future__ import annotations
 from dataclasses import dataclass, replace
 from typing import TYPE_CHECKING, Any, Protocol
 
-from mindroom.agents import remove_run_by_event_id
 from mindroom.coalescing import coalesced_prompt
 from mindroom.conversation_resolver import MessageContext
-from mindroom.conversation_state_writer import (
-    LoadPersistedTurnMetadataRequest,
-    PersistedTurnMetadata,
-    RemoveStaleRunsRequest,
-)
-from mindroom.handled_turns import HandledTurnLedger, HandledTurnRecord, HandledTurnState
+from mindroom.handled_turns import HandledTurnState
 from mindroom.hooks.ingress import hook_ingress_policy
 from mindroom.matrix.identity import extract_agent_name
 from mindroom.matrix.message_content import extract_edit_body
-from mindroom.post_response_effects import matrix_run_metadata_for_handled_turn, record_handled_turn
-from mindroom.thread_utils import should_agent_respond
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
@@ -29,13 +21,12 @@ if TYPE_CHECKING:
     from mindroom.bot_runtime_view import BotRuntimeView
     from mindroom.constants import RuntimePaths
     from mindroom.conversation_resolver import ConversationResolver
-    from mindroom.conversation_state_writer import ConversationStateWriter
-    from mindroom.dispatch_planner import IngressHookRunner
     from mindroom.hooks import MessageEnvelope
     from mindroom.matrix.client import ResolvedVisibleMessage
     from mindroom.matrix.event_info import EventInfo
     from mindroom.message_target import MessageTarget
-    from mindroom.tool_system.runtime_context import ToolRuntimeSupport
+    from mindroom.turn_policy import IngressHookRunner
+    from mindroom.turn_store import TurnStore
 
 
 class _GenerateResponse(Protocol):
@@ -69,11 +60,9 @@ class EditRegeneratorDeps:
     get_logger: Callable[[], structlog.stdlib.BoundLogger]
     runtime_paths: RuntimePaths
     agent_name: str
-    get_handled_turn_ledger: Callable[[], HandledTurnLedger]
     resolver: ConversationResolver
-    state_writer: ConversationStateWriter
-    tool_runtime: ToolRuntimeSupport
-    dispatch_hook_service: IngressHookRunner
+    turn_store: TurnStore
+    ingress_hook_runner: IngressHookRunner
     generate_response: _GenerateResponse
 
 
@@ -86,9 +75,6 @@ class EditRegenerator:
     def _logger(self) -> structlog.stdlib.BoundLogger:
         return self.deps.get_logger()
 
-    def _handled_turn_ledger(self) -> HandledTurnLedger:
-        return self.deps.get_handled_turn_ledger()
-
     def _client(self) -> nio.AsyncClient:
         client = self.deps.runtime.client
         if client is None:
@@ -98,27 +84,7 @@ class EditRegenerator:
 
     def _mark_source_events_responded(self, handled_turn: HandledTurnState) -> None:
         """Mark one or more source events as handled by the same response."""
-        record_handled_turn(self._handled_turn_ledger(), handled_turn)
-
-    def load_persisted_turn_metadata(
-        self,
-        *,
-        room: nio.MatrixRoom,
-        thread_id: str | None,
-        original_event_id: str,
-        requester_user_id: str,
-    ) -> PersistedTurnMetadata | None:
-        """Load persisted run metadata for one edited turn when available."""
-        return self.deps.state_writer.load_persisted_turn_metadata(
-            LoadPersistedTurnMetadataRequest(
-                room=room,
-                thread_id=thread_id,
-                original_event_id=original_event_id,
-                requester_user_id=requester_user_id,
-            ),
-            build_message_target=self.deps.resolver.build_message_target,
-            build_tool_execution_identity=self.deps.tool_runtime.build_execution_identity,
-        )
+        self.deps.turn_store.mark_handled(handled_turn)
 
     async def edit_regeneration_context(
         self,
@@ -147,57 +113,6 @@ class EditRegenerator:
             requires_full_thread_history=context.requires_full_thread_history,
         )
 
-    def remove_stale_runs_for_turn_record(
-        self,
-        *,
-        turn_record: HandledTurnRecord,
-        recorded_turn_context_available: bool,
-        room: nio.MatrixRoom,
-        thread_id: str | None,
-        original_event_id: str,
-        requester_user_id: str,
-    ) -> None:
-        """Remove stale persisted runs using the recorded turn context when possible."""
-        if (
-            recorded_turn_context_available
-            and turn_record.conversation_target is not None
-            and turn_record.history_scope is not None
-        ):
-            self.deps.state_writer.remove_stale_runs_for_turn_record(
-                turn_record=turn_record,
-                requester_user_id=requester_user_id,
-                build_tool_execution_identity=self.deps.tool_runtime.build_execution_identity,
-                remove_run_by_event_id_fn=remove_run_by_event_id,
-            )
-            return
-        self.remove_stale_runs_for_edited_message(
-            room=room,
-            thread_id=thread_id,
-            original_event_id=original_event_id,
-            requester_user_id=requester_user_id,
-        )
-
-    def remove_stale_runs_for_edited_message(
-        self,
-        *,
-        room: nio.MatrixRoom,
-        thread_id: str | None,
-        original_event_id: str,
-        requester_user_id: str,
-    ) -> None:
-        """Remove persisted runs tied to the pre-edit message before regenerating."""
-        self.deps.state_writer.remove_stale_runs_for_edited_message(
-            RemoveStaleRunsRequest(
-                room=room,
-                thread_id=thread_id,
-                original_event_id=original_event_id,
-                requester_user_id=requester_user_id,
-            ),
-            build_message_target=self.deps.resolver.build_message_target,
-            build_tool_execution_identity=self.deps.tool_runtime.build_execution_identity,
-            remove_run_by_event_id_fn=remove_run_by_event_id,
-        )
-
     async def handle_message_edit(  # noqa: C901, PLR0911, PLR0912, PLR0915
         self,
         room: nio.MatrixRoom,
@@ -216,42 +131,27 @@ class EditRegenerator:
             self._logger().debug("ignoring_edit_from_other_agent", agent=sender_agent_name)
             return
 
-        turn_record = self._handled_turn_ledger().get_turn_record(original_event_id)
-        context = await self.edit_regeneration_context(
-            room,
-            event,
-            conversation_target=turn_record.conversation_target if turn_record is not None else None,
-        )
-        persisted_turn_metadata = self.load_persisted_turn_metadata(
+        loaded_turn = self.deps.turn_store.load_turn_record(
             room=room,
-            thread_id=context.thread_id,
+            thread_id=event_info.thread_id or event_info.thread_id_from_edit,
             original_event_id=original_event_id,
             requester_user_id=requester_user_id,
         )
-        if turn_record is None:
-            if persisted_turn_metadata is None:
-                self._logger().debug(
-                    "No handled turn record found for edited message",
-                    original_event_id=original_event_id,
-                )
-                return
-            turn_record = HandledTurnRecord(
-                anchor_event_id=persisted_turn_metadata.anchor_event_id,
-                source_event_ids=persisted_turn_metadata.source_event_ids,
-                response_event_id=persisted_turn_metadata.response_event_id,
-                source_event_prompts=persisted_turn_metadata.source_event_prompts,
+        if loaded_turn is None:
+            self._logger().debug(
+                "No handled turn record found for edited message",
+                original_event_id=original_event_id,
             )
-        recorded_turn_context_available = bool(
-            turn_record.conversation_target is not None and turn_record.history_scope is not None,
+            return
+        turn_record = loaded_turn.record
+        context = await self.edit_regeneration_context(
+            room,
+            event,
+            conversation_target=turn_record.conversation_target,
         )
-        response_owner_missing = turn_record.response_owner is None
-        if response_owner_missing and persisted_turn_metadata is not None:
+        if loaded_turn.response_owner_missing:
             turn_record = replace(turn_record, response_owner=self.deps.agent_name)
-        response_event_id = (
-            persisted_turn_metadata.response_event_id
-            if persisted_turn_metadata is not None and persisted_turn_metadata.response_event_id is not None
-            else turn_record.response_event_id
-        )
+        response_event_id = turn_record.response_event_id
         if response_event_id is None:
             self._logger().debug("missing_previous_response_for_edit", event_id=original_event_id)
             return
@@ -260,7 +160,7 @@ class EditRegenerator:
             thread_id=context.thread_id,
             reply_to_event_id=turn_record.anchor_event_id,
         )
-        regeneration_history_scope = turn_record.history_scope or self.deps.state_writer.history_scope()
+        regeneration_history_scope = turn_record.history_scope or self.deps.turn_store.default_history_scope()
         regeneration_response_owner = turn_record.response_owner or self.deps.agent_name
         if regeneration_response_owner != self.deps.agent_name:
             self._logger().debug(
@@ -270,24 +170,12 @@ class EditRegenerator:
             )
             return
         needs_turn_record_backfill = (
-            turn_record.response_event_id != response_event_id
-            or response_owner_missing
+            loaded_turn.requires_backfill
+            or loaded_turn.response_owner_missing
             or turn_record.history_scope is None
             or turn_record.conversation_target is None
-            or (
-                turn_record.is_coalesced
-                and turn_record.source_event_prompts is None
-                and persisted_turn_metadata is not None
-                and persisted_turn_metadata.source_event_prompts is not None
-            )
         )
         coalesced_source_event_prompts = turn_record.source_event_prompts
-        if (
-            coalesced_source_event_prompts is None
-            and persisted_turn_metadata is not None
-            and persisted_turn_metadata.is_coalesced
-        ):
-            coalesced_source_event_prompts = persisted_turn_metadata.source_event_prompts
 
         self._logger().info(
             "Regenerating response for edited message",
@@ -345,7 +233,7 @@ class EditRegenerator:
                 conversation_target=regeneration_target,
             )
             regeneration_turn_record = replace(regeneration_turn_record, source_event_prompts=updated_prompt_map)
-            regeneration_matrix_run_metadata = matrix_run_metadata_for_handled_turn(regeneration_handled_turn)
+            regeneration_matrix_run_metadata = self.deps.turn_store.matrix_run_metadata(regeneration_handled_turn)
         else:
             regeneration_prompt = edited_content
             regeneration_matrix_run_metadata = None
@@ -359,32 +247,13 @@ class EditRegenerator:
             source_kind="edit",
         )
         ingress_policy = hook_ingress_policy(envelope)
-        if await self.deps.dispatch_hook_service.emit_message_received_hooks(
+        if await self.deps.ingress_hook_runner.emit_message_received_hooks(
             envelope=envelope,
             correlation_id=event.event_id,
             policy=ingress_policy,
         ):
             self._mark_source_events_responded(regeneration_handled_turn)
             return
-
-        if turn_record.response_owner is None:
-            should_respond = should_agent_respond(
-                agent_name=self.deps.agent_name,
-                am_i_mentioned=context.am_i_mentioned,
-                is_thread=context.is_thread,
-                room=room,
-                thread_history=context.thread_history,
-                config=self.deps.runtime.config,
-                runtime_paths=self.deps.runtime_paths,
-                mentioned_agents=context.mentioned_agents,
-                has_non_agent_mentions=context.has_non_agent_mentions,
-                sender_id=requester_user_id,
-            )
-            if not should_respond and not regeneration_turn_record.is_coalesced:
-                self._logger().debug("Agent should not respond to edited message")
-                if needs_turn_record_backfill:
-                    self._mark_source_events_responded(regeneration_handled_turn)
-                return
 
         regenerated_event_id = await self.deps.generate_response(
             room_id=room.room_id,
@@ -399,9 +268,11 @@ class EditRegenerator:
             response_envelope=envelope,
             correlation_id=event.event_id,
             matrix_run_metadata=regeneration_matrix_run_metadata,
-            on_lifecycle_lock_acquired=lambda: self.remove_stale_runs_for_turn_record(
-                turn_record=regeneration_turn_record,
-                recorded_turn_context_available=recorded_turn_context_available,
+            on_lifecycle_lock_acquired=lambda: self.deps.turn_store.remove_stale_runs_for_edit(
+                loaded_turn=replace(
+                    loaded_turn,
+                    record=regeneration_turn_record,
+                ),
                 room=room,
                 thread_id=context.thread_id,
                 original_event_id=original_event_id,
