+++ tests/test_edit_event_handling.py
@@ -106,7 +106,6 @@ async def test_bot_ignores_edit_events(tmp_path: Path) -> None:
         ),
         patch.object(bot._turn_controller, "_dispatch_text_message", new_callable=AsyncMock) as mock_dispatch,
         patch.object(bot._edit_regenerator, "handle_message_edit", new_callable=AsyncMock) as mock_handle_edit,
-        patch.object(bot._edit_regenerator, "load_persisted_turn_metadata", return_value=None),
     ):
         # Process the edit event - this should not re-enter normal dispatch.
         await bot._on_message(room, edit_event)
@@ -204,7 +203,6 @@ async def test_bot_ignores_multiple_edits(tmp_path: Path) -> None:
         ),
         patch.object(bot._turn_controller, "_dispatch_text_message", new_callable=AsyncMock) as mock_dispatch,
         patch.object(bot._edit_regenerator, "handle_message_edit", new_callable=AsyncMock) as mock_handle_edit,
-        patch.object(bot._edit_regenerator, "load_persisted_turn_metadata", return_value=None),
     ):
         # Process all edit events
         for edit_event in edit_events:
@@ -292,7 +290,7 @@ async def test_regular_agent_ignores_edits(tmp_path: Path) -> None:
     # Mock the generate_response method
     with (
         patch.object(bot, "_generate_response", new_callable=AsyncMock) as mock_generate,
-        patch.object(bot._edit_regenerator, "load_persisted_turn_metadata", return_value=None),
+        patch.object(bot._turn_store, "load_turn_record", return_value=None),
     ):
         # Process the edit event
         await bot._on_message(room, edit_event)
