+++ tests/test_preformed_team_routing.py
@@ -129,7 +129,7 @@ async def test_router_does_not_route_when_preformed_team_is_mentioned(config_wit
     event = _mock_event_with_team_mention(team_user_id)
 
     # Also patch suggest_agent_for_message to detect accidental routing
-    with patch("mindroom.dispatch_planner.suggest_agent_for_message", new=AsyncMock(return_value="a1")):
+    with patch("mindroom.turn_controller.suggest_agent_for_message", new=AsyncMock(return_value="a1")):
         await router._on_message(room, event)
 
     # Router must not send any message (i.e., must not route)
