+++ tests/test_routing.py
@@ -300,10 +300,10 @@ class TestAIRouting:
         mock_event = MagicMock()
         mock_event.body = "Test message"
 
-        with patch("mindroom.dispatch_planner.suggest_agent_for_message") as mock_suggest:
+        with patch("mindroom.turn_controller.suggest_agent_for_message") as mock_suggest:
             # Should raise AssertionError since general is not the router agent
             with pytest.raises(AssertionError):
-                await bot._dispatch_planner.execute_router_relay(
+                await bot._turn_controller._execute_router_relay(
                     mock_room,
                     mock_event,
                     [],
